Level Prototype without proper Lava collision and placement
This is probably the penultimate python file that I will upload. The next one should have all the basic elements in it.

I got rid of nearly all magic numbers. The ones that remain mostly relate to the hero's size and arbitrary placement of pipes and what not.

Collision for lava is going to be very similar to that of pipes. I am not sure whether I want "daleks" to die when they touch lava, but if they were, I would probably copy the way they collide with pipes.